fox two cases of format strings != arguments on 32-bit OS

+++ dbdimp.c
@@ -2069,7 +2069,7 @@ int ib_blob_write(SV *sth, imp_sth_t *imp_sth, XSQLVAR *var, SV *value)
     p = string;
     while (total_length > 0)
     {
-        DBI_TRACE_imp_xxh(imp_sth, 3, (DBIc_LOGPIO(imp_sth), "ib_blob_write: %"PRIdMAX" bytes left\n", total_length));
+        DBI_TRACE_imp_xxh(imp_sth, 3, (DBIc_LOGPIO(imp_sth), "ib_blob_write: %lld bytes left\n", (long long)total_length));
 
         /* set new segment start pointer */
         seg = p;
